GUILD-522: Convert schema_class_mixin to a module class & apply base class fixes
