+++ lib/deimos/active_record_producer.rb
@@ -58,9 +58,9 @@ module Deimos
         payload.delete_if do |k, _|
           k.to_sym != :payload_key && !fields.map(&:name).include?(k)
         end
-        return payload unless use_schema_classes?(config.to_h)
+        return payload unless Utils::SchemaClass.use_schema_classes?(config.to_h)
 
-        schema_class_instance(payload, config[:schema])
+        Utils::SchemaClass.schema_class_instance(payload, config[:schema])
       end
 
       # Query to use when polling the database with the DbPoller. Add
