+++ lib/generators/deimos/schema_class/templates/schema_record.rb.tt
@@ -38,7 +38,7 @@
 <% end -%>
     # @override
     <%= @initialization_definition %>
-      super()
+      super
 <%- @fields.each do |field| -%>
       self.<%= field.name %> = <%= field.name %>
 <% end -%>
